// src/game/gameplay/ui/enums/enums.ts
export enum EButtonKey {
  START_BUTTON = "start_button",
  LINK_BUTTON = "link_button",
  NEXT_DAY = "next_day",
  PLANTS_BUTTON = "plants_button",
  ANIMALS_BUTTON = "animals_button",
  CONSTRUCTION_BUTTON = "construction_button",
  STRAWBERRY_BUTTON = "strawberry_button",
  CORN_BUTTON = "corn_button",
  TOMATO_BUTTON = "tomato_button",
  GRAPE_BUTTON = "grape_button",
  CHICKEN_BUTTON = "chicken_button",
  SHEEP_BUTTON = "sheep_button",
  COW_BUTTON = "cow_button",
  FENCE_BUTTON = "fence_button",
  BARN_BUTTON = "barn_button"
}

export enum EButtonGroupKey {
  ITEM_TYPE = "item_type",
  PLANTS = "plants",
  ANIMALS = "animals",
  CONSTRUCTION = "construction"
}